Module ai_kit::planner

The planner module implements a basic system for backtracking planner.

It supports features like:

specifying a goal with constraints that must be satisfied by the resultant plan.

the ability to solve a conjunction of goal

rendering of a plan in graphviz format for easier visualization

Structs

ConjunctivePlanner

Goal

Planner

PlanningConfig

Enums

InvalidPlan

UnificationIndex

Functions

first_goal_to_increment
Determine the first goal to increment